Remove unsupported download formats, use consumer code
Removes parquet and duckdb download end points from the dataset
controller.  This only affects the publishing side.  Switched to using
the consumer side code for downloads.  This allows us to remove the
download code from the cube service when the new cube builder is
switched in.
